sched/cpufreq: zenith: widen rate_limit_us * NSEC_PER_USEC to u64
up_rate_limit_us and down_rate_limit_us are stored as unsigned
int and multiplied by NSEC_PER_USEC (1000) before being written
to *_rate_delay_ns (s64). The multiplication was performed in
u32 width:

    z_pol->up_rate_delay_ns = val * NSEC_PER_USEC;

which overflows when val > 4,294,967 (roughly 4.29 s in usec),
silently truncating the result and feeding a nonsensical
negative s64 to the rate-limit comparisons. Same pattern exists
in down_rate_limit_us_store() and in zenith_start() where the
initial derived delay is computed.

Cast the left operand to u64 so the multiply happens in 64-bit
and wraps only at ~1.8e16 ns (~213 days), well beyond anything
a sane user would write. No change in behaviour for reasonable
values; just stops the overflow in the pathological one.

+++ kernel/sched/cpufreq_zenith.c
@@ -1796,7 +1796,7 @@ static ssize_t up_rate_limit_us_store(struct gov_attr_set *attr_set, const char 
 	t->up_rate_limit_us = val;
 
 	list_for_each_entry(z_pol, &attr_set->policy_list, tunables_hook) {
-		z_pol->up_rate_delay_ns = val * NSEC_PER_USEC;
+		z_pol->up_rate_delay_ns = (u64)val * NSEC_PER_USEC;
 		update_min_rate_limit_ns(z_pol);
 	}
 	return count;
@@ -1818,7 +1818,7 @@ static ssize_t down_rate_limit_us_store(struct gov_attr_set *attr_set, const cha
 	t->down_rate_limit_us = val;
 
 	list_for_each_entry(z_pol, &attr_set->policy_list, tunables_hook) {
-		z_pol->down_rate_delay_ns = val * NSEC_PER_USEC;
+		z_pol->down_rate_delay_ns = (u64)val * NSEC_PER_USEC;
 		update_min_rate_limit_ns(z_pol);
 	}
 	return count;
@@ -2058,8 +2058,8 @@ static int zenith_start(struct cpufreq_policy *policy)
 	struct zenith_policy *z_policy = policy->governor_data;
 	unsigned int cpu;
 
-	z_policy->up_rate_delay_ns = z_policy->tunables->up_rate_limit_us * NSEC_PER_USEC;
-	z_policy->down_rate_delay_ns = z_policy->tunables->down_rate_limit_us * NSEC_PER_USEC;
+	z_policy->up_rate_delay_ns = (u64)z_policy->tunables->up_rate_limit_us * NSEC_PER_USEC;
+	z_policy->down_rate_delay_ns = (u64)z_policy->tunables->down_rate_limit_us * NSEC_PER_USEC;
 	update_min_rate_limit_ns(z_policy);
 
 	z_policy->last_freq_update_time = 0;
